+++ scrooge-core/src/main/scala/com/twitter/scrooge/ThriftUnion.scala
@@ -1,5 +1,7 @@
 package com.twitter.scrooge
 
+import com.twitter.util.Memoize
+
 /**
  * Unions are tagged with this trait.
  */
@@ -27,4 +29,28 @@ trait ThriftUnion extends ThriftStruct {
    * Returns `None` if this represents an unknown union field.
    */
   def unionStructFieldInfo: Option[ThriftStructFieldInfo]
+}
+
+object ThriftUnion {
+  private[this] val fieldInfos = Memoize.classValue { c =>
+    if (!classOf[ThriftUnion].isAssignableFrom(c)) {
+      throw new IllegalArgumentException(s"${c.getName} does not extend ThriftUnion")
+    }
+    ThriftStructMetaData
+      .forStructClass(c.asInstanceOf[Class[ThriftUnion]])
+      .unionFields
+      .collectFirst {
+        case f if f.fieldClassTag.runtimeClass == c => f.structFieldInfo
+      }
+  }
+
+  /**
+   * Given a specific union member class, returns the struct field info for it.
+   * It returns the same value as invoking `unionStructFieldInfo` on an instance
+   * of the class, without having to create an instance of it.
+   * @param c the union member class
+   * @return the struct field info for the class, or None
+   */
+  def fieldInfoForUnionClass(c: Class[_ <: ThriftUnion]): Option[ThriftStructFieldInfo] =
+    fieldInfos(c)
 }